Project/AutoCodeDocsAI/requirements.txt
google-adk
GitPython